Some organization, plus animated lock POC
Moved all initial views into a `first-time` folder, to get the `app` folder to start resembling the actual app hierarchy more.

Changed the bottom toggle switch into an animated lock icon.

Used a combination of normal css (for class transitions) and inline React element styles, which lets me use logic and variables on the styles, and keep them inline with the rest of the component.

A lot of the benefits feel lost by having to do both, but I do really like describing an element in one place.  Did it this way partly to start a discussion, and try out what this style might look like.