Refactor Vault Generation and Restoring
Not as deep a refactor as I'd like, but a lot more than I meant to do in a single commit...

 - 12 word phrase now persists until it has been confirmed.
 - Creating vault now supports entering random entropy text.

Fixed some nondeterministic view rendering. Or at least it seemed nondeterminsitic. Like we shouldn't have a `currentView` property if we aren't using it to derive the current view. We still have a slight mess of overrides during initialization, but it's slightly better now.

 - Added a global loading indicator for use wherever it is needed (not styled yet) state.appState.isLoading.
 - Added a `app/template.js` file for easily creating new components in our current format.

Deleted a lot of cruft from the `first-time/init` component, where it still had legacy copies of several views that were since copied to other files.